blog-gen/requirements.txt
sentence-transformers
uvicorn
ctransformers
ipykernel
langchain
python-box
streamlit
